+++ tests/parser/charsetFilteredOutputStreamTest.cpp
@@ -101,10 +101,10 @@ VMIME_TEST_SUITE_BEGIN(charsetFilteredOutputStreamTest)
 			std::ostringstream testName;
 			testName << i << ": " << entry.fromCharset << " -> " << entry.toCharset;
 
-			const unsigned int inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
+			const unsigned long inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
 			vmime::string in(entry.fromBytes, entry.fromBytes + inLength);
 
-			const unsigned int outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
+			const unsigned long outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
 			vmime::string expectedOut(entry.toBytes, entry.toBytes + outLength);
 
 			vmime::string actualOut;
@@ -136,10 +136,10 @@ VMIME_TEST_SUITE_BEGIN(charsetFilteredOutputStreamTest)
 			std::ostringstream testName;
 			testName << i << ": " << entry.fromCharset << " -> " << entry.toCharset;
 
-			const unsigned int inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
+			const unsigned long inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
 			vmime::string in(entry.fromBytes, entry.fromBytes + inLength);
 
-			const unsigned int outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
+			const unsigned long outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
 			vmime::string expectedOut(entry.toBytes, entry.toBytes + outLength);
 
 			vmime::string actualOut;
@@ -174,10 +174,10 @@ VMIME_TEST_SUITE_BEGIN(charsetFilteredOutputStreamTest)
 			std::ostringstream testName;
 			testName << i << ": " << entry.fromCharset << " -> " << entry.toCharset;
 
-			const unsigned int inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
+			const unsigned long inLength = (entry.fromLength == 0 ? strlen(entry.fromBytes) : entry.fromLength);
 			vmime::string in(entry.fromBytes, entry.fromBytes + inLength);
 
-			const unsigned int outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
+			const unsigned long outLength = (entry.toLength == 0 ? strlen(entry.toBytes) : entry.toLength);
 			vmime::string expectedOut(entry.toBytes, entry.toBytes + outLength);
 
 			vmime::string actualOut;
